requirements.txt
discord.py
gspread
flask
oauth2client